refactor(tokens): rename rule
