--- packages/eslint-plugin-pf-codemods/src/rules/v6/tokensWarn/tokens-warn.md
@@ -1,17 +1,0 @@
-### tokens-warn
-
-We have updated our CSS tokens. About half of our tokens have been replaced with newer ones. To find a suitable replacement token, check our [v6 token documentation](https://staging-v6.patternfly.org/tokens/all-patternfly-tokens).
-
-#### Examples
-
-In:
-
-```jsx
-%inputExample%
-```
-
-Out:
-
-```jsx
-%outputExample%
-```
